Updated E2E CI scripts and Tinybird sync flow
ref https://linear.app/ghost/issue/BER-3363/unified-environment-manager

Moved CI shell logic into versioned scripts, parallelized build-mode runtime prep, and made infra startup sync/reset Tinybird state.
