+++ e2e/package.json
@@ -13,9 +13,11 @@
     "build:docker": "docker build -f Dockerfile.e2e --build-arg GHOST_IMAGE=${GHOST_E2E_BASE_IMAGE:?Set GHOST_E2E_BASE_IMAGE} -t ${GHOST_E2E_IMAGE:-ghost-e2e:local} ..",
     "prepare": "tsc --noEmit",
     "pretest": "(test -n \"$GHOST_E2E_SKIP_BUILD\" || test -n \"$CI\") && echo 'Skipping Docker build' || echo 'Tip: run yarn dev or yarn workspace @tryghost/e2e infra:up before running tests'",
-    "infra:up": "docker compose -f ../compose.dev.yaml -f ../compose.dev.analytics.yaml up -d --wait mysql redis mailpit tinybird-local tb-cli analytics",
-    "infra:down": "docker compose -f ../compose.dev.yaml -f ../compose.dev.analytics.yaml stop analytics tb-cli tinybird-local mailpit redis mysql",
-    "tinybird:fetch-config": "node ./scripts/fetch-tinybird-config.mjs",
+    "infra:up": "bash ./scripts/infra-up.sh",
+    "infra:down": "bash ./scripts/infra-down.sh",
+    "tinybird:sync": "node ./scripts/sync-tinybird-state.mjs",
+    "tinybird:fetch-config": "yarn tinybird:sync",
+    "preflight:build": "bash ./scripts/prepare-ci-e2e-build-mode.sh",
     "test": "playwright test --project=main",
     "test:analytics": "playwright test --project=analytics",
     "test:all": "playwright test --project=main --project=analytics",
